Split importer to separate concerns
